Update Assignment.html
Finally added the functions to add names into the queue. I also added a function to delete the first person from the queue. I also used bootstrap in order to format the html such that the queue isn't in an unusual area. This is the first working version of the code.